pyproject.toml
[tool.poetry]
name = "rag4p"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nltk = "^3.8.1"
openai = "^1.8.0"
pycryptodome = "^3.20.0"
requests = "^2.31.0"
configparser = "^6.0.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.4"
scipy = "^1.11.4"
onnxruntime = "1.16.3"
tokenizers = "^0.15.0"
unidecode = "^1.3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
